Immutable lexer, bug fixes
