+++ src/Kingsland.MofParser/Ast/LiteralValueArrayAst.cs
@@ -13,6 +13,12 @@ namespace Kingsland.MofParser.Ast
 
         public sealed class Builder
         {
+
+            public Builder()
+            {
+                this.Values = new List<LiteralValueAst>();
+            }
+
             public List<LiteralValueAst> Values
             {
                 get;
