+++ src/Kingsland.MofParser/Ast/MethodDeclarationAst.cs
@@ -14,6 +14,11 @@ namespace Kingsland.MofParser.Ast
         public sealed class Builder
         {
 
+            public Builder()
+            {
+                this.Parameters = new List<ParameterDeclarationAst>();
+            }
+
             public IdentifierToken Name
             {
                 get;
